Work on the display code.
I need an LCD and a OLED display with I2C on the Pico. The LCD largely is what the example showed in the documentation.

The OLED version is work in progress. There is an Arduino version "SSD1306Ascii" that I used before. Let's take this as a base.

DON'T FORGET TO PUT IN THE COPYRIGHT of the creators.